fix for travis failing due to local scope issue

+++ modules/assimp/import_utils.h
@@ -269,7 +269,7 @@ public:
 		paths.push_back(p_path.get_base_dir().plus_file("../texture/" + path.get_file().get_basename() + extension));
 		paths.push_back(p_path.get_base_dir().plus_file("../texture/" + path.get_file() + extension));
 		paths.push_back(p_path.get_base_dir().plus_file("../texture/" + path.get_file()));
-		for (size_t i = 0; i < paths.size(); i++) {
+		for (int i = 0; i < paths.size(); i++) {
 			if (dir.file_exists(paths[i])) {
 				found = true;
 				path = paths[i];
